+++ packages/cta-admin/src/registries/list-options/import.tsx
@@ -34,7 +34,7 @@ export const ImportListOption = (): JSX.Element => {
 
 						<p style={ { margin: 0 } }>
 							{ __(
-								'Popup Maker Pro gives you the power to import your call to actions from a JSON file in seconds.',
+								'Popup Maker Pro gives you the power to import your calls to action from a JSON file in seconds.',
 								'popup-maker'
 							) }
 						</p>